Change page name back
