--- src/js/common/schemaform/pages/RoutedFormPage.jsx
@@ -1,173 +1,0 @@
-import PropTypes from 'prop-types';
-import React from 'react';
-import { connect } from 'react-redux';
-import { withRouter } from 'react-router';
-import Scroll from 'react-scroll';
-import _ from 'lodash/fp';
-
-import SchemaForm from '../SchemaForm';
-import ProgressButton from '../../components/form-elements/ProgressButton';
-import { setData, uploadFile } from '../actions';
-import { getNextPage, getPreviousPage } from '../routing';
-
-import { focusElement } from '../../utils/helpers';
-
-function focusForm() {
-  const legend = document.querySelector('.form-panel legend:not(.schemaform-label)');
-  if (legend && legend.getBoundingClientRect().height > 0) {
-    focusElement(legend);
-  } else {
-    focusElement('.nav-header');
-  }
-}
-
-const scroller = Scroll.scroller;
-const scrollToTop = () => {
-  scroller.scrollTo('topScrollElement', window.VetsGov.scroll || {
-    duration: 500,
-    delay: 0,
-    smooth: true,
-  });
-};
-
-class RoutedFormPage extends React.Component {
-  componentDidMount() {
-    if (!this.props.blockScrollOnMount) {
-      scrollToTop();
-      focusForm();
-    }
-  }
-
-  componentDidUpdate(prevProps) {
-    if (this.props.route.pageKey !== prevProps.route.pageKey
-      || prevProps.params.index !== this.props.params.index) {
-      scrollToTop();
-      focusForm();
-    }
-  }
-
-  onChange(formData) {
-    let newData = formData;
-    if (this.props.route.pageConfig.showPagePerItem) {
-      // If this is a per item page, the formData object will have data for a particular
-      // row in an array, so we need to update the full form data object and then call setData
-      newData = _.set([this.props.route.pageConfig.arrayPath, this.props.params.index], formData, this.props.form.data);
-    }
-    this.props.setData(newData);
-  }
-
-  onSubmit({ formData }) {
-    const { form, params, route, location } = this.props;
-
-    // This makes sure defaulted data on a page with no changes is saved
-    // Probably safe to do this for regular pages, too, but it hasn’t been necessary
-    if (route.pageConfig.showPagePerItem) {
-      const newData = _.set([route.pageConfig.arrayPath, params.index], formData, form.data);
-      this.props.setData(newData);
-    }
-
-    const page = getNextPage(route.pageList, form.data, location.pathname, route.pageConfig);
-
-    this.props.router.push(page);
-  }
-
-  goBack = () => {
-    const { form, route: { pageConfig, pageList }, location } = this.props;
-    const page = getPreviousPage(pageList, form.data, location.pathname, pageConfig);
-
-    this.props.router.push(page);
-  }
-
-  render() {
-    const {
-      route,
-      params,
-      form,
-      contentAfterButtons,
-      prefilled
-    } = this.props;
-
-    let {
-      schema,
-      uiSchema
-    } = form.pages[route.pageConfig.pageKey];
-
-    let data = form.data;
-
-    if (route.pageConfig.showPagePerItem) {
-      const arrayPath  = route.pageConfig.arrayPath;
-      // Instead of passing through the schema/uiSchema to SchemaForm, the
-      // current item schema for the array at arrayPath is pulled out of the page state and passed
-      schema = schema.properties[arrayPath].items[params.index];
-      // Similarly, the items uiSchema and the data for just that particular item are passed
-      uiSchema = uiSchema[arrayPath].items;
-      // And the data should be for just the item in the array
-      data = _.get([arrayPath, params.index], data);
-    }
-
-    return (
-      <div className="form-panel">
-        <SchemaForm
-          name={route.pageConfig.pageKey}
-          title={route.pageConfig.title}
-          data={data}
-          schema={schema}
-          uiSchema={uiSchema}
-          pagePerItemIndex={params ? params.index : undefined}
-          prefilled={prefilled}
-          uploadFile={this.props.uploadFile}
-          onChange={this.props.setData}
-          onSubmit={this.onSubmit}>
-          <div className="row form-progress-buttons schemaform-buttons">
-            <div className="small-6 medium-5 columns">
-              <ProgressButton
-                onButtonClick={this.goBack}
-                buttonText="Back"
-                buttonClass="usa-button-secondary"
-                beforeText="«"/>
-            </div>
-            <div className="small-6 medium-5 end columns">
-              <ProgressButton
-                submitButton
-                buttonText="Continue"
-                buttonClass="usa-button-primary"
-                afterText="»"/>
-            </div>
-          </div>
-          {contentAfterButtons}
-        </SchemaForm>
-      </div>
-    );
-  }
-}
-
-function mapStateToProps(state) {
-  return {
-    form: state.form,
-    user: state.user
-  };
-}
-
-const mapDispatchToProps = {
-  setData,
-  uploadFile
-};
-
-RoutedFormPage.propTypes = {
-  form: PropTypes.object.isRequired,
-  route: PropTypes.shape({
-    pageConfig: PropTypes.shape({
-      pageKey: PropTypes.string.isRequired,
-      schema: PropTypes.object.isRequired,
-      uiSchema: PropTypes.object.isRequired
-    }),
-    pageList: PropTypes.arrayOf(PropTypes.shape({
-      path: PropTypes.string.isRequired
-    }))
-  }),
-  setData: PropTypes.func
-};
-
-export default withRouter(connect(mapStateToProps, mapDispatchToProps)(RoutedFormPage));
-
-export { RoutedFormPage };
